+++ src/api/editor/api.ts
@@ -3,8 +3,6 @@ import { ModuleProfile, Api, API } from "../../types"
 import { HighlightPosition } from './type'
 
 export interface IEditorApi extends Api {
-  events: {
-  }
   highlight(position: HighlightPosition, filePath: string, hexColor: string): void
   discardHighlight(): void
 }
@@ -20,6 +18,8 @@ export abstract class EditorApi extends BaseApi<IEditorApi> implements API<IEdit
     super(localProfile)
   }
 
+  /** Remove the current highlight if any, and add a new one */
   abstract highlight(position: HighlightPosition, filePath: string, hexColor: string): void
+  /** Remove the current hightlight if any */
   abstract discardHighlight(): void
 }
